kv-proxy/requirements.txt
fastapi
uvicorn
python-dotenv
httpx